requirements.txt
streamlit==1.25.0
ultralytics==8.3.49
torch==2.1.0
pandas
pillow
matplotlib